+++ files/urls.py
@@ -7,7 +7,6 @@ from . import management_views, views
 urlpatterns = [
     re_path(r"^$", views.index),
     re_path(r"^about", views.about, name="about"),
-    re_path(r"^add_subtitle", views.add_subtitle, name="add_subtitle"),
     re_path(r"^categories$", views.categories, name="categories"),
     re_path(r"^contact$", views.contact, name="contact"),
     re_path(r"^edit", views.edit_media, name="edit_media"),