+++ files/views.py
@@ -29,7 +29,7 @@ from cms.custom_pagination import FastPaginationWithoutCount
 from cms.permissions import IsAuthorizedToAdd, IsSuperUser, IsUserOrEditor, user_allowed_to_upload
 from users.models import User
 
-from .forms import ContactForm, MediaForm, SubtitleForm
+from .forms import ContactForm, MediaForm
 from .helpers import clean_query, produce_ffmpeg_commands
 from .methods import (
     get_user_or_session,
@@ -73,31 +73,6 @@ def about(request):
 
     context = {}
     return render(request, "cms/about.html", context)
-
-
-@login_required
-def add_subtitle(request):
-    """Add subtitle view"""
-
-    friendly_token = request.GET.get("m", "").strip()
-    if not friendly_token:
-        return HttpResponseRedirect("/")
-    media = Media.objects.filter(friendly_token=friendly_token).first()
-    if not media:
-        return HttpResponseRedirect("/")
-
-    if not (request.user == media.user or is_mediacms_editor(request.user) or is_mediacms_manager(request.user)):
-        return HttpResponseRedirect("/")
-
-    if request.method == "POST":
-        form = SubtitleForm(media, request.POST, request.FILES)
-        if form.is_valid():
-            subtitle = form.save()
-            messages.add_message(request, messages.INFO, "Subtitle was added!")
-            return HttpResponseRedirect(subtitle.media.get_absolute_url())
-    else:
-        form = SubtitleForm(media_item=media)
-    return render(request, "cms/add_subtitle.html", {"form": form})
 
 
 def categories(request):
@@ -196,7 +171,7 @@ def edit_media(request):
     return render(
         request,
         "cms/edit_media.html",
-        {"form": form, "add_subtitle_url": media.add_subtitle_url},
+        {"form": form},
     )