+++ frontend/src/static/js/components/media-page/ViewerInfoContent.js
@@ -80,20 +80,6 @@ function EditMediaButton(props) {
   );
 }
 
-function EditSubtitleButton(props) {
-  let link = props.link;
-
-  if (window.MediaCMS.site.devEnv) {
-    link = '#';
-  }
-
-  return (
-    <a href={link} rel="nofollow" title="Edit subtitle" className="edit-subtitle">
-      EDIT SUBTITLE
-    </a>
-  );
-}
-
 export default function ViewerInfoContent(props) {
   const { userCan } = useUser();
 
@@ -203,14 +189,9 @@ export default function ViewerInfoContent(props) {
             />
           ) : null}
 
-          {userCan.editMedia || userCan.editSubtitle || userCan.deleteMedia ? (
+          {userCan.editMedia || userCan.deleteMedia ? (
             <div className="media-author-actions">
               {userCan.editMedia ? <EditMediaButton link={MediaPageStore.get('media-data').edit_url} /> : null}
-              {userCan.editSubtitle && 'video' === MediaPageStore.get('media-data').media_type ? (
-                <EditSubtitleButton
-                  link={MediaPageStore.get('media-data').edit_url.replace('edit?', 'add_subtitle?')}
-                />
-              ) : null}
 
               <PopupTrigger contentRef={popupContentRef}>
                 <button className="remove-media">DELETE MEDIA</button>